Add Subreddit.__eq__ to fix == comparisons.
This fixes several bugs that were using the default (identity
comparison) to check if two subreddits were the same. Sometimes
the two subreddit objects referred to the same subreddit but
weren't the same instance.

Fixes empty NSFW listings for non-logged users and the subreddit
top bar not highlighting current subreddits.