Update regexp for authorize duplicate customer error
Fixes the following bug:
A user was unable to authorize campaign because clicking Pay raised
an exception.

Cause:
The user's record was missing from the authorize_account_id table. Our code
has some logic to parse the customer id from the authorize error response when
this happens and rewrite the record, but the regexp was out of date.

This change updates the regexp to match the current authorize.net error msg
format and logs some info to make it easier to catch the problem if the format
changes again in the future.

Note:
- Why would a user record be missing in the first place? Should be looked into.