DenormalizedRelation: Add in LastModified support.
I did an experiment to see if DenormalizedRelation would need the
LastModified stuff, since it takes heavy advantage of row-level bloom
filters. It turns out that there are a huge number of completely
not-found rows being fetched per second that LastModified would save us
from doing a round trip for.

Since LastModified is shared between several relations in a single
request, and CommentVote, Save, and Hide will be denormalized soon, this
should give us a good amount of savings in the long run.