Use isinstance(x, basestring) in wrap_links()
For some data structures (specifically, LiveAdWeights), json.loads()
is used to de-serialize the data. json.dumps/loads returns unicode
strings, not byte strings. While it would be more appropriate for
this data to be in a byte string, it is non-trivial to configure
the json module to do so systematically.