plugins: Ensure load_controllers is called in paster shell/run as well.
load_controllers is the safe place to do imports of things that rely on
`g` et al being in the correct state. It was not being called in paster
shell or run which meant that any hooks defined in a module that relies
on `g` at module scope (e.g. things importing model or controller code)
would not be connected in those contexts.