REVERT ME: Temporary stuff to disentangle commentstree_q.
The rollout of sharded commentstree_q stuff involves splitting the
new_comment routing key from the commentstree_q bind. Until all code is
in place, the bind will continue to exist though. This means that
commentstree_q will receive messages that are being handled by one of
commentstree_*_q and likely slow things down further. To avoid this, we
temporarily look for an application header to know that new code sent
the new_comment message to this queue and therefor we can safely ignore
the message.