Enable fetching recommendations for multiple srs at once.
This change allows you to pass a list of srid36s and get back recommendations
for all of them at once, which is useful for multireddit suggestions.

A function that accepted a single subreddit id as an argument and memoized the
results was removed and its functionality replaced by sgm caching, which sets
multiple items in memcache at once.

A source prefix is added to the row key so multiple sets of recs can be stored
in the same column family.

Adult items are filtered unless one of the original subreddits was over_18.

Recommendations for a set of subreddits are merged and sorted by the number of
items matched. (User feedback showed that people perceive recommendations as
incorrect if they don't match the "theme" of a large multi, and recs that
match several srs in the multi instead of just one are more likely to fit the
theme, so we show those first.)