/dev/api: Document the use of c.site.
For POST requests, this is done (preferably) by the "r" parameter. Other
requests use the path prefix /r/{subreddit}. Since c.site is a global,
it's hard to actually see if it's used by methods, so this uses an
additional manually added parameter to the api_doc decorator to indicate
when c.site is important.