setup.py: Include a dependency link for PyCAPTCHA 0.4.
PyPI only has versions for Python 2.3 and 2.4. Modern installs that
don't use our PPA to get a debianized copy of this library fail when
trying to resolve the setup.py install_requires dependencies. This fixes
the problem by using a reddit-hosted tarball of the 0.4 source (taken
directly from the source SVN repo).