SRMember: Start dual-writing to a dedicated cache pool.
This is one of the most heavily hit rels currently and the hope is that
pulling it out to its own pool will make the main cache more effective
for everything else.  It may make sense to move SRMember to a different
model such as a C* DenormalizedRelation in the future.