CSS Filter: replace cssutils with tinycss2.
The aim of this patch is to allow a mostly equivalent subset of CSS
through as the cssutils-based parser did.

The subset of valid properties represented in this are the entirety of
CSS2.1 plus a handful of newer ones that are already in active use on
reddit.

Several IE hacks like the "filter" property, "_height", etc. are no
longer allowed.