timetext: Undo logic flip that broke times >1 day.
During the move to core, timetext's chunk logic was flipped around as an
optimization. This didn't actually work, though that's been unnoticeable
in production because the maxage perfectly masked it.

For a timestamp 32 hours ago, "1 day ago" would be the expected output,
however "32 hours ago" would be displayed instead. This is because the
chunk values are in seconds while they're compared against "count" which
is the current unit-of-interest (hours, days, months, etc.)

This reverts that logic to the original form which was how the python
code works.